requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-migrate
flask-jwt-extended
passlib
requests
rq
redis